ui: Refactor event history to use links + add go to child workflow
