+++ frontend/src/client/types.gen.ts
@@ -379,6 +379,7 @@ export type EventGroup = {
     retry_policy?: ActionRetryPolicy;
     start_delay?: number;
     join_strategy?: JoinStrategy;
+    related_wf_exec_id?: string | null;
 };
 
 export type EventHistoryResponse = {
